+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 ### Changes
+* Dropped support for Python 3.5.
+* Opening empty files raises ValueError.
+* Using an invalid index type to select a cell now raises a ValueError.
 
 ### Fixes